e2e helm-test: dependency as file (#352)
